reforma-terminada
arquivos transferidos para o novo modelo com sucesso, arquivos da pasta pages apagados
