+++ magmartfront/src/app/admin/pokemon/[id]/page.tsx
@@ -2,14 +2,9 @@
 import React,{ useEffect, useState } from 'react';
 import styles from '@/styles/Product.module.css';
 import "@/app/globals.css";
-import Image from 'next/image'
-import { useRouter } from 'next/router';
 import AdmNavbar from '@/components/AdmNavbar';
-import Card from '@/components/Card';
 import Footer from '@/components/Footer';
-import PropTypes from 'prop-types';
 import GetPokemon from '@/APIs/getPokemon';
-import { types } from 'util';
 import { Button } from '@mui/material';
 import DeletePokemon from '@/APIs/deletePokemon'; 
 import Link from 'next/link';
@@ -18,7 +13,6 @@ import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
 import Select from '@mui/material/Select';
-import { stringify } from 'querystring';
 
 function Product({params}: {params: {id: string}}) {
     const [pok_id, setPokId] = useState('');
@@ -39,27 +33,24 @@ function Product({params}: {params: {id: string}}) {
     useEffect(() => {
         async function getdados() {
             try {
-                //setPokId(queryString);
-                //if (typeof router.query.data === 'string') {
-                    data = await GetPokemon(params.id);
-                    setPokId(data.id);
-                    setRefId(data.reference_id);
-                    setName(data.name);
-                    setWeight(data.weight);
-                    setHeight(data.height);
-                    setExp(data.experience);
-                    setPrice(data.price);
-                    setStock(data.in_stock);
-                    setImage(data.image_url);
-                    setType((data.types.reduce((types, type)=>{
-                        return types.concat(` ${type.name}`)
-                    },"")));
-                    setTierName(data.tier.name);
-                    setTierId(data.tier.id);
-                    setMinExp(data.tier.minimal_experience);
-                    setLimitExp(data.tier.limit_experience);
-                    console.log(data)
-                //}
+                data = await GetPokemon(params.id);
+                setPokId(data.id);
+                setRefId(data.reference_id);
+                setName(data.name);
+                setWeight(data.weight);
+                setHeight(data.height);
+                setExp(data.experience);
+                setPrice(data.price);
+                setStock(data.in_stock);
+                setImage(data.image_url);
+                setType((data.types.reduce((types, type)=>{
+                    return types.concat(` ${type.name}`)
+                },"")));
+                setTierName(data.tier.name);
+                setTierId(data.tier.id);
+                setMinExp(data.tier.minimal_experience);
+                setLimitExp(data.tier.limit_experience);
+                console.log(data)
                 
             } catch {
                 console.log('erro')
