+++ magmartfront/src/app/admin/pokemon/create/page.tsx
@@ -8,7 +8,7 @@ import Box from '@mui/material/Box';
 import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
-import Select, { SelectChangeEvent } from '@mui/material/Select';
+import Select from '@mui/material/Select';
 
 export default function CadPokemon() {
     const [tierId, setTierId] = useState('');
