+++ magmartfront/src/app/home/page.tsx
@@ -1,11 +1,8 @@
 "use client"
 import React,{ useState, useEffect } from 'react';
 import Link from 'next/link';
-import styles from '@/styles/Home.module.css';
 import "@/app/globals.css";
-import Image from 'next/image'
 import Navbar from '@/components/Navbar';
-import Card from '@/components/Card';
 import Footer from '@/components/Footer';
 import GetAllPokemon from '@/APIs/getAllPokemon';
 import { type } from 'os';
